Move session helper in api/v1
